fixed provinces PST GST and HST columns

+++ db/migrate/20200412183958_create_provinces.rb
@@ -2,9 +2,9 @@ class CreateProvinces < ActiveRecord::Migration[6.0]
   def change
     create_table :provinces do |t|
       t.string :name
-      t.integer :PST
-      t.integer :HST
-      t.integer :GST
+      t.decimal :PST
+      t.decimal :HST
+      t.decimal :GST
       
       t.timestamps
     end
